Add files via upload
At first I downloaded a csv file of a list of suicide rates but I could not access the codebook for that data set as it was a condensed data set.
This meant that I could not properly do the coding assignment but in class I downloaded the given data sample from schoology and then I was able to apply the skills we learnt in class to siphon through the data and find a piece of data that I am interested in which was what type of family people were living with. This is important to me as I am interested to compare the data sets from the USA to the data sets from the United Kingdom. potato. One of the more interesting things I have discovered is that there are significantly more people living in a non family household in the USA than in the United Kingdom which could be because that there is less people that are living in the UK but also because there are less same sex couples that live in the UK (percentage) which would be the people that primarily fit into this group.